requirements.txt
deepagents>=0.4.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.4.0
langsmith>=0.3.0
ddgs>=9.0.0
streamlit>=1.30.0
python-dotenv>=1.0.0
